feat: parser migrate to tools
