+++ packages/mirage-engine/dev/index.html
@@ -124,7 +124,7 @@
             align-items: center;
             justify-content: center;
             border-radius: 30px;
-            background: linear-gradient(to right, red, blue);
+            background: linear-gradient(in oklab to right, red, blue);
           "
         >
           <span>test, test, test</span>
